+++ yak_extensions/lib/yak_extensions.dart
@@ -1,4 +1,0 @@
-library yak_extensions;
-
-export 'src/iterable/iterable_extensions.dart';
-export 'src/num/num_extensions.dart';