+++ yak_runner/lib/src/shared/void/type_void.dart
@@ -1,0 +1,5 @@
+/// if you are wondering why this nonsense
+/// [https://stackoverflow.com/a/65671199/9836706]
+
+Type _type<T>() => T;
+final Type typeVoid = _type<void>();