+++ yak_test/lib/src/yak_runner/extension/tryarg/void/delegate_test.dart
@@ -1,7 +1,0 @@
-import 'package:yak_runner/yak_runner.dart';
-import 'package:yak_test/src/yak_runner/factory/all.dart';
-
-extension TryRunTryArgTestFactoryExtension on TryRunTryArg {
-  TryRunTryArgTestFactory buildTestFactory(String description) =>
-      TryRunTryArgTestFactory(description);
-}