+++ yak_test/lib/src/yak_runner/extension/trytry/async_delegate_test.dart
@@ -1,7 +1,0 @@
-import 'package:yak_runner/yak_runner.dart';
-import 'package:yak_test/src/yak_runner/factory/all.dart';
-
-extension TryRunTryAsyncTestFactoryExtension on TryRunTryAsync {
-  TryRunTryAsyncTestFactory buildTestFactory(String description) =>
-      TryRunTryAsyncTestFactory(description);
-}