+++ yak_test/lib/src/yak_runner/factory/async/nullary.dart
@@ -1,0 +1,10 @@
+import 'package:yak_runner/yak_runner.dart';
+import '../../test/all.dart';
+
+class YakRunnerAsyncTestFactory implements YakRunnerTestFactoryDelegate {
+  const YakRunnerAsyncTestFactory(this.description);
+  final String description;
+  @override
+  YakRunnerTestDelegate<T> call<T>() =>
+      YakRunnerAsyncTest<T>(description: description);
+}