+++ yak_test/lib/src/yak_runner/factory/simple/void/delegate_test.dart
@@ -1,9 +1,0 @@
-import 'package:yak_runner/yak_runner.dart';
-import '../../../testers/simple/void/delegate_test.dart';
-
-class TryRunTestFactory implements VoidTestFactoryDelegate {
-  const TryRunTestFactory(this.description);
-  final String description;
-  @override
-  VoidTestDelegate call() => TryRunTester(description: description);
-}