+++ yak_test/lib/src/yak_runner/test/all.dart
@@ -1,2 +1,2 @@
-export 'sync/all.dart';
 export 'async/all.dart';
+export 'sync/all.dart';