+++ yak_test/lib/yak_test.dart
@@ -1,4 +1,3 @@
 library yak_test;
 
-export 'src/yak_runner/extension/all.dart';
-export 'src/yak_runner/factory/all.dart';
+export 'src/all.dart';