+++ yak_test/test/yak_runner/test/async/nullary_test.dart
@@ -1,0 +1,7 @@
+import 'package:yak_test/src/yak_runner/all.dart';
+
+void main() {
+  final _tester =
+      YakRunnerAsyncTest<int>(description: '`YakRunnerTest<int>` sample test');
+  _tester(4);
+}