+++ yak_test/test/yak_runner/testers/arg/any/delegate_tester_.dart
@@ -1,8 +1,0 @@
-import 'package:yak_test/src/yak_runner/testers/all.dart';
-
-void main() {
-  const _tester = TryAnyRunArgTester<int, String>(
-    description: '`TryAnyRunArgTester<int, String>` sample test',
-  );
-  _tester(4, 'hello');
-}