+++ yak_test/test/yak_runner/testers/arg/void/async_delegate_tester_test.dart
@@ -1,8 +1,0 @@
-import 'package:yak_test/src/yak_runner/testers/all.dart';
-
-void main() {
-  const _tester = TryRunArgAsyncTester<int>(
-    description: '`TryRunArgAsyncTester<int>` sample test',
-  );
-  _tester(4);
-}