+++ yak_test/test/yak_runner/testers/tryarg/any/async_delegate_tester_test.dart
@@ -1,8 +1,0 @@
-import 'package:yak_test/src/yak_runner/testers/all.dart';
-
-void main() {
-  const _tester = TryAnyRunTryArgAsyncTester<int, String>(
-    description: '`TryAnyRunTryArgAsyncTester<int, String>` sample test',
-  );
-  _tester(4, 'hello');
-}