+++ yak_test/test/yak_runner/testers/tryarg/void/async_delegate_tester_test.dart
@@ -1,8 +1,0 @@
-import 'package:yak_test/src/yak_runner/testers/all.dart';
-
-void main() {
-  const _tester = TryRunTryArgAsyncTester<String>(
-    description: '`TryRunTryArgTester<String>` sample test',
-  );
-  _tester('hello');
-}